+++ tutorials/conversion_examples/example_opendrive_to_commonroad.py
@@ -8,8 +8,10 @@ from crdesigner.common.config.opendrive_config import open_drive_config
 from crdesigner.common.file_writer import CRDesignerFileWriter, OverwriteExistingFile
 from crdesigner.map_conversion.map_conversion_interface import opendrive_to_commonroad
 
-input_path = Path.cwd().parent.parent / "tests/map_conversion/test_maps/opendrive/opendrive-1.xodr"
-output_path = Path.cwd() / "example_files/opendrive/opendrive-1.xml"
+#input_path = Path("/home/sebastian/Downloads/Birmingham_sample_1218.xodr/Birmingham_sample_1218.xodr")
+input_path = Path.cwd().parent / "conversion_examples/with_bug.xodr"
+
+output_path = Path.cwd() / "example_files/opendrive/birmingham.xml"
 
 config = open_drive_config
 config.lanelet_types_backwards_compatible = False
@@ -17,20 +19,12 @@ config.lanelet_types_backwards_compatible = False
 # load OpenDRIVE file, parse it, and convert it to a CommonRoad scenario
 scenario = opendrive_to_commonroad(input_path)
 
-# store converted file as CommonRoad scenario
-writer = CRDesignerFileWriter(
-    scenario=scenario,
-    planning_problem_set=PlanningProblemSet(),
-    author="Sebastian Maierhofer",
-    affiliation="Technical University of Munich",
-    source="CommonRoad Scenario Designer",
-    tags={Tag.URBAN},
-)
+import matplotlib.pyplot as plt
+from commonroad.visualization.mp_renderer import MPRenderer
 
-# create a folder for the example file if it does not exist
-if os.path.exists(Path.cwd() / "example_files") is False:
-    os.mkdir(Path.cwd() / "example_files")
-if os.path.exists(Path.cwd() / "example_files/opendrive") is False:
-    os.mkdir(Path.cwd() / "example_files/opendrive")
-
-writer.write_to_file(str(output_path), OverwriteExistingFile.ALWAYS)
+# plot the planning problem and the scenario for the fifth time step
+plt.figure(figsize=(25, 10))
+rnd = MPRenderer()
+scenario.draw(rnd)
+rnd.render()
+plt.show()